refactor: change domain logic of game start
- added GameControl class to control game
- moved a bridge making function to GameContol
- changed readBridgeSize function in InputView
- deleted Bridge class
- changed file name PrintConstant to PrintMessage to ensure the function
- changed BridgeSize directory and logic to distinguish MVC Model
