+++ src/controller/BridgeGame.js
@@ -1,7 +1,20 @@
+const { makeBridge } = require('../BridgeMaker');
+const { generate } = require('../BridgeRandomNumberGenerator');
+
 /**
  * 다리 건너기 게임을 관리하는 클래스
  */
 class BridgeGame {
+  #bridge;
+
+  constructor() {
+    this.#bridge = [];
+  }
+
+  make(size) {
+    this.#bridge = makeBridge(size, generate);
+  }
+
   /**
    * 사용자가 칸을 이동할 때 사용하는 메서드
    * <p>
