--- src/model/Bridge.js
@@ -1,9 +1,0 @@
-class Bridge {
-  #bridge;
-
-  constructor(bridge) {
-    this.#bridge = bridge;
-  }
-}
-
-module.exports = Bridge;
