+++ src/view/InputView.js
@@ -1,6 +1,9 @@
 const MissionUtils = require('@woowacourse/mission-utils');
-const Message = require('../constant/PrintConstant');
-const BridgeSize = require('../controller/BridgeSize');
+// const { makeBridge } = require('../BridgeMaker');
+const Message = require('../constant/PrintMessage');
+// const BridgeGame = require('../controller/BridgeGame');
+const BridgeSize = require('../model/BridgeSize');
+const Direction = require('../model/Direction');
 
 /**
  * 사용자로부터 입력을 받는 역할을 한다.
@@ -10,6 +13,7 @@ const InputView = {
    * 다리의 길이를 입력받는다.
    */
   readBridgeSize() {
+    let size = 0;
     MissionUtils.Console.readLine(Message.INPUT_LENGTH, answer => {
       try {
         const bridgeSize = new BridgeSize(Number(answer));
@@ -17,7 +21,9 @@ const InputView = {
         MissionUtils.Console.print(error.message);
         this.readBridgeSize();
       }
+      size = answer;
     });
+    return size;
   },
 
   /**
